+++ src/components/multi-dashboard/PlayerBalances.tsx
@@ -1,5 +1,5 @@
 import { PlayerCardCollectionData } from "@/types/playerCardCollection";
-import { SplBalance } from "@/types/spl/balances";
+import { PlayerPoolBalances, SplBalance } from "@/types/spl/balances";
 import { SPLSeasonRewards } from "@/types/spl/seasonRewards";
 import { Box } from "@mui/material";
 import CardCollection from "./CardCollection";
@@ -11,27 +11,29 @@ import TopBalances from "./TopBalances";
 import ElectroneumBalances from "@/components/multi-dashboard/ElectroneumBalances";
 
 interface Props {
-  username: string;
   balances?: SplBalance[];
+  poolBalances?: PlayerPoolBalances;
   seasonRewards?: SPLSeasonRewards;
   glintLoading?: boolean;
   glintError?: string | null;
   collectionData?: PlayerCardCollectionData | null;
   collectionLoading?: boolean;
+  collectionError?: string | null;
 }
 
 export default function PlayerBalances({
-  username,
   balances,
+  poolBalances,
   seasonRewards,
   glintLoading,
   glintError,
   collectionData,
   collectionLoading,
+  collectionError,
 }: Readonly<Props>) {
   return (
     <Box sx={{ display: "flex", flexDirection: "row", flexWrap: "wrap", gap: 1 }}>
-      <TopBalances balances={balances} />
+      <TopBalances balances={balances} poolBalances={poolBalances} />
       <ElectroneumBalances balances={balances} />
 
       <Box sx={{ display: "flex", gap: 1, width: "100%" }}>
@@ -46,11 +48,7 @@ export default function PlayerBalances({
           glintLoading={glintLoading}
           glintError={glintError}
         />
-        <CardCollection
-          username={username}
-          externalData={collectionData}
-          externalLoading={collectionLoading}
-        />
+        <CardCollection data={collectionData} loading={collectionLoading} error={collectionError} />
       </Box>
     </Box>
   );